# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tabaq | طبق - Food & Dining Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatusResponse"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatusResponse"
  /users/{userId}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      summary: Get user's followers
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCard"
  /users/{userId}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      summary: Get users this user follows
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of following
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCard"
  /users/{userId}/reviews:
    get:
      operationId: getUserReviews
      tags: [users]
      summary: Get reviews written by a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /users/{userId}/bookings:
    get:
      operationId: getUserBookings
      tags: [users]
      summary: Get bookings for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, past, cancelled]
      responses:
        "200":
          description: User bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /users/{userId}/leaderboard-rank:
    get:
      operationId: getUserLeaderboardRank
      tags: [users]
      summary: Get user's points, level, and rank
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard rank info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardRank"
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [users]
      summary: Get top users leaderboard
      parameters:
        - name: cityId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Leaderboard list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # ─── Search ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    # ─── Geo ──────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        email:
          type: string
        nameEn:
          type: string
        nameAr:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        isVerified:
          type: boolean
        points:
          type: integer
        level:
          type: integer
        levelTitle:
          type: string
        preferredLanguage:
          type: string
          enum: [en, ar]
        cityId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - points
        - level
        - levelTitle
        - isVerified
        - preferredLanguage
        - createdAt
    UserCard:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        avatarUrl:
          type: string
        isVerified:
          type: boolean
        level:
          type: integer
        levelTitle:
          type: string
        reviewCount:
          type: integer
      required:
        - id
        - isVerified
        - level
        - levelTitle
        - reviewCount
    UserProfile:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        reviewCount:
          type: integer
        bookingCount:
          type: integer
        followerCount:
          type: integer
        followingCount:
          type: integer
        isFollowing:
          type: boolean
      required:
        - user
        - reviewCount
        - bookingCount
        - followerCount
        - followingCount
        - isFollowing
    CreateUserRequest:
      type: object
      properties:
        phone:
          type: string
        email:
          type: string
        nameEn:
          type: string
        nameAr:
          type: string
        preferredLanguage:
          type: string
          enum: [en, ar]
          default: en
        cityId:
          type: integer
      required:
        - preferredLanguage
    UpdateUserRequest:
      type: object
      properties:
        nameEn:
          type: string
        nameAr:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        preferredLanguage:
          type: string
          enum: [en, ar]
        cityId:
          type: integer
    LeaderboardRank:
      type: object
      properties:
        userId:
          type: integer
        points:
          type: integer
        level:
          type: integer
        levelTitle:
          type: string
        rank:
          type: integer
        reviewCount:
          type: integer
        nextLevelPoints:
          type: integer
      required:
        - userId
        - points
        - level
        - levelTitle
        - rank
        - reviewCount
        - nextLevelPoints
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        user:
          $ref: "#/components/schemas/UserCard"
        points:
          type: integer
        reviewCount:
          type: integer
      required:
        - rank
        - user
        - points
        - reviewCount

    # ─── Restaurants ──────────────────────────────────────────────────────
...
    FollowStatusResponse:
      type: object
      properties:
        isFollowing:
          type: boolean
        followerCount:
          type: integer
      required:
        - isFollowing
        - followerCount

    # ─── Menus ────────────────────────────────────────────────────────────
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        restaurantId:
          type: integer
        restaurantNameEn:
          type: string
        restaurantNameAr:
          type: string
        restaurantCoverImageUrl:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        partySize:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed, no_show]
        occasionId:
          type: integer
        specialRequests:
          type: string
        referenceCode:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - restaurantId
        - restaurantNameEn
        - restaurantNameAr
        - date
        - time
        - partySize
        - status
        - referenceCode
        - createdAt
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userNameEn:
          type: string
        userNameAr:
          type: string
        userAvatarUrl:
          type: string
        userLevel:
          type: integer
        userLevelTitle:
          type: string
        restaurantId:
          type: integer
        dishId:
          type: integer
        ratingOverall:
          type: number
          minimum: 1
          maximum: 5
        ratingFood:
          type: number
          minimum: 1
          maximum: 5
        ratingService:
          type: number
          minimum: 1
          maximum: 5
        ratingAmbiance:
          type: number
          minimum: 1
          maximum: 5
        ratingValue:
          type: number
          minimum: 1
          maximum: 5
        textEn:
          type: string
        textAr:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        isLiked:
          type: boolean
        visitDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userNameEn
        - ratingOverall
        - likeCount
        - isLiked
        - createdAt
...