# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tabaq | طبق - Food & Dining Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferredLanguage } from "./userPreferredLanguage";

export interface User {
  id: number;
  phone?: string;
  email?: string;
  nameEn?: string;
  nameAr?: string;
  avatarUrl?: string;
  bio?: string;
  isVerified: boolean;
  points: number;
  level: number;
  levelTitle: string;
  preferredLanguage: UserPreferredLanguage;
  cityId?: number;
  createdAt: Date;
}